% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMGT_Vab_Hsap.R
\docType{data}
\name{ab.hsap}
\alias{ab.hsap}
\title{Homo sapiens V-alpha and V-beta sequences from IMGT}
\format{
An object of class \code{"data.frame"}.
}
\usage{
data(IMGT_Vab_Hsap)
}
\description{
Homo sapiens V-alpha and V-beta sequences from IMGT
}
\examples{
data(IMGT_Vab_Hsap)
\donttest{

}
}
\keyword{datasets}
