% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TCRdist}
\alias{TCRdist}
\alias{cdrDist}
\alias{test}
\title{TCRdist}
\usage{
cdrDist(cdrx, cdry, cdr)

test(cdrx, cdry)
}
\arguments{
\item{cdr1, }{matrix}
}
\description{



}
\author{
Julian Spagnuolo

Julian Spagnuolo
}
