% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{trimmer}
\alias{trimmer}
\title{trimmer}
\usage{
trimmer(cdr3, vpart, jpart)
}
\description{
trims V and J sequences from CDR3 amino acid sequences based on the IMGT reference.
}
\author{
Julian Spagnuolo
}
