% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Vdist}
\alias{Vdist}
\title{Vdist}
\usage{
Vdist(cdrs, Vgene, gap_penalty, weight)
}
\description{
compute pair-wise V distances using IMGT CDR1, CDR2 and CDR2.5 alignments
}
\author{
Julian Spagnuolo
}
